Extract `WalletConnect` to own file
